Split overlay tests into multiple files
